+++ src/mongodb/IdentifiableMongoDbPersistence.ts
@@ -5,6 +5,7 @@ import { ConfigParams } from 'pip-services-commons-node';
 import { FilterParams } from 'pip-services-commons-node';
 import { PagingParams } from 'pip-services-commons-node';
 import { DataPage } from 'pip-services-commons-node';
+import { AnyValueMap } from 'pip-services-commons-node';
 import { IIdentifiable } from 'pip-services-commons-node';
 import { IStringIdentifiable } from 'pip-services-commons-node';
 import { ObjectWriter } from 'pip-services-commons-node';
@@ -181,14 +182,40 @@ export class IdentifiableMongoDbPersistence<T extends IIdentifiable<K>, K> exten
             return;
         }
 
-        let newItem = _.omit(item.id);
+        let newItem = _.omit(item, 'id');
         let options = {
             new: true
         };
 
-        this._model.findByIdAndUpdate(item.id, item, options, (err, newItem) => {
+        this._model.findByIdAndUpdate(item.id, newItem, options, (err, newItem) => {
             if (!err)
                 this._logger.trace(correlationId, "Update in %s with id = %s", this._collection, item.id);
+
+            if (callback) {
+                newItem = this.convertToPublic(newItem);
+                callback(err, newItem);
+            }
+        });
+    }
+
+    public updatePartially(correlationId: string, id: K, data: AnyValueMap,
+        callback?: (err: any, item: T) => void): void {
+            
+        if (data == null || id == null) {
+            if (callback) callback(null, null);
+            return;
+        }
+
+        let newItem = {
+            $set: data.getAsObject()
+        };
+        let options = {
+            new: true
+        };
+
+        this._model.findByIdAndUpdate(id, newItem, options, (err, newItem) => {
+            if (!err)
+                this._logger.trace(correlationId, "Update partially in %s with id = %s", this._collection, id);
 
             if (callback) {
                 newItem = this.convertToPublic(newItem);